requirements.txt
requests
beautifulsoup4
pandas
supabase
dotenv
pymupdf
pdfplumber
camelot-py[cv]
pytesseract
pandas
Pillow
sentence-transformers
numpy

